// src/components/settings/types.ts
export type ThemePreference = "light" | "dark" | "system";

export type FacebookTokenValues = {
  eaag: string;
  eaab: string;
  eaai: string;
  eaah: string;
  cookie: string;
};